Hardcode weight for instructions with  MultiAssetFilter params (#822)
* Hardcode weight for instructions with  MultiAssetFilter params


* TEst for receiver weight


* Add sanity weight checks for typical sender and receiver xcm messages


* Fix clippy


* Fix clippy

+++ pallets/manta-pay/src/lib.rs
@@ -549,7 +549,7 @@ pub mod pallet {
                 }
             }
             (
-                VoidNumberSetInsertionOrder::<T>::contains_key(max_sender_index as u64),
+                VoidNumberSetInsertionOrder::<T>::contains_key(max_sender_index),
                 senders,
             )
         }
